requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
hindsight-all>=0.1.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-multipart>=0.0.6

